tests: move `transfer_funcs` into `util`
